// console/src/generated/QuotaAllowance.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type QuotaAllowance = { used: string | null, limit: string | null, remaining: string | null, used_percent: string | null, unlimited: boolean, unit: string | null, period_start: number | null, period_end: number | null, };